/home/repa/haskell/bayes_classifier/bayesClassifier.hs autocommit from emacs

+++ bayes_classifier/bayesClassifier.hs
@@ -63,7 +63,7 @@ main = do
       
       [trainData'] = computeP $ traverse likelyhoods' (\(Z :. i :. j) -> (Z :. i :. j))
                                                       (\lp (Z :. i :. j) -> (fromIntegral (lp (Z :. i :. j))) / 
-                                                                             fromIntegral (sums ! (Z :. i)))
+                                                                             (sums ! (Z :. i)))
 
 --      likelyhoods is append after I sum across all indexes -> Array U (Z :. labelIdx) (Array U (Z :. WordIdx) Int)
